sutvarkiau form1 funkcionaluma
