+++ holy-water/Baras.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace holy_water
+{
+    public class Baras
+        {
+            public String name;
+            public double volume;
+            public int percentage;
+            public double locX, locY;
+        }
+}
